Fixes #1 (line 1180 -- duplicate T w/ missing ;)
The original line is:

```
1180 PRINT "AFTER ALL YOUR PURCHASES, YOU NOW HAVE ";T;T" DOLLARS LEFT"
```

It seems like there should be only one T, surrounded by semicolons.
However, if there are to be two Ts, all items to be printed need to be
delimited by semicolons.

Also, OS/8 BASIC automatically adds spaces around the items, so the
trailing and leading spaces in the strings have been removed.